Support OCI 1.0
SOCI relies on being able to associate an index with an image. This
capability was introduced in OCI distribution and image specs version
1.1 as Artifacts and the Referrers API respectively.

In order to drive adoption, an OCI 1.0 compatible fallback was added
that encodes artifacts as Image Manifests and stores the Referrers API
content in an Image Index.

This PR adds the ability to serialize and deserialize SOCI indices as
either OCI Artifacts or OCI Images. There is no automatic OCI 1.0
fallback when creating images. The user of the CLI/library must
explicitly request an OCI 1.0 compatible Image Manifest. The snapshotter
will be able to automatically load a SOCI index in either serialized
form.

Signed-off-by: Kern Walster <walster@amazon.com>