Skip updating grpc with bump-deps
Containerd communicates with SOCI over grpc. In order for that
communication to work, both SOCI and the containerd library must be
using compatible versions of GRPC. A recent update (1.56.2 -> 1.57.0)
is not backwards compatible and SOCI does not build.

To work around this issue so that we can continue to update other
packages, this change prevents updating GRPC with the bump-deps script.

Signed-off-by: Kern Walster <walster@amazon.com>