Skip grpc in bump-deps.sh
The latest version of grpc is not compatible with the version containerd
is using. Skip automatic updates.

This change also stops using the `-u` flag in `go get`. The flag causes
`go get` to update transitive depdendencies which means that grpc was
getting update whenever we tried to update containerd.

Signed-off-by: Kern Walster <walster@amazon.com>